All Refactoring
